+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/MyTestWithWorkflowsApplication.java
@@ -11,7 +11,7 @@
 limitations under the License.
 */
 
-package io.dapr.it.spring.boot.testcontainers;
+package io.dapr.it.testcontainers;
 
 import io.dapr.testcontainers.DaprModule;
 import org.springframework.boot.SpringApplication;
