+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/TestWorkflow.java
@@ -11,7 +11,7 @@
 limitations under the License.
 */
 
-package io.dapr.it.spring.boot.testcontainers;
+package io.dapr.it.testcontainers;
 
 import io.dapr.workflows.Workflow;
 import io.dapr.workflows.WorkflowStub;
